+++ src/Symfony/Component/Workflow/Tests/Subject.php
@@ -5,10 +5,12 @@ namespace Symfony\Component\Workflow\Tests;
 final class Subject
 {
     private $marking;
+    private $context;
 
     public function __construct($marking = null)
     {
         $this->marking = $marking;
+        $this->context = [];
     }
 
     public function getMarking()
@@ -16,8 +18,14 @@ final class Subject
         return $this->marking;
     }
 
-    public function setMarking($marking)
+    public function setMarking($marking, array $context = [])
     {
         $this->marking = $marking;
+        $this->context = $context;
+    }
+
+    public function getContext(): array
+    {
+        return $this->context;
     }
 }